Fix possible memory corruption in `tp_st`
Currently, `tp_st` passes in an out_buffer to a Rust callback. If the
callback succeeds, tp_st returns Ok(out_buffer); otherwise if a YDBError
occurs, it reuses the out_buffer to return an error; otherwise the
callback failed and returned an error: tp_st returns that error.

The first problem which I found Saturday is that if a user error occurs,
it will mostly likely write that error into out_buffer - `tp` does this
in the context_api module. In this case, out_buffer will be freed twice:
once when `tp_st` returns and twice whenever the error is dropped. My
solution to this was to always call mem::forget on the out_buffer when a
user error occurs - this can waste memory but it prevents a double free.

The second problem I found today is that the user callback can consume
the out_buffer even if an error does not occur! This is because the type
for the callback is `FnMut(u64, Vec<u8>) -> Result<Vec<u8>, Box<dyn
Error>>`: the output buffer does not have to be the same as the input
buffer. The first solution I thought of is to pass in a &mut Vec
instead: `FnMut(u64, &mut Vec<u8>) -> Result<(), Box<dyn Error>>` - that
will prevent the Rust callback from dropping the buffer at all. This has
the drawback that the buffer cannot be reused to store errors, but will
never result in a double-free or a memory leak.

The final problem I thought of is that if the callback gets passed a
&mut Vec, then it can change any part of the vec, including calling (the
Rust equivalent of) realloc. These changes will not be seen by the
original out_buffer (to pass it to C, I have to make a copy of the {
buf_addr, len_used, len_alloc } metadata, although the elements do not
have to be copied). There are a few possible solutions to this:
      - return the buffer in callback_struct.return_value. However, then
        the function could still free the buffer, which would cause a
        use-after-free if YDB then returns an engine error (such as
        YDB_ERR_TPTIMEOUT).
      - pass in &mut[u8], which allows changing the elements in the
        buffer, but not reallocating or freeing the buffer. The downsides here
        are that the Rust function would have to allocate a new buffer to return
        an error or call the other functions in Key::* and the context_api can
        no longer reuse the `errstr` passed in for `context.buffer` (since it is
        not allowed to reallocate).