Add test that `Context` can't be used with async
If used in an async context, multiple functions could change each
others' tptoken, leading to strange and hard-to-debug errors from
yottadb.

This is not caught by the compiler since we are using `RefCell`.
This is not caught by `RefCell` since the changes are not necessarily
happening at the same time; the faulty pattern looks like this:

task1: enter tp_st; call `await` on some other task
task2: exit tp_st (resets tptoken to YDB_NOTTP)
task1: set or get the value for a Key (WRONG! Uses YDB_NOTTP instead
of the transaction token)

This is not caught by `impl !Sync for RefCell` since tokio only requires
`Send`.

Ideally this would just be `impl !Send for Context {}`, but negative
trait bounds aren't stable.

I'm not sure if there's runtimes besides tokio which don't require
`Send`, but hopefully this as well as the documentation will get the
point across that async + Context is A BAD IDEA.