Add test for multiple keys in ydb_lock
This uncovered a few issues in the API.
I changed it to take `&[Key]` instead of `&[&Key]` which makes more
sense if you have a `Vec<Key>`.
This does make `ydb_lock` a little more annoying to call with only one
argument: you have to use `slice::from_ref`, which I documented.