[#31] Don't pre-allocate context buffer
This avoids wasting memory that's never used.

This commit also removes many unnecessary uses of `with_capacity` in
doc-tests and examples.

 ## Background

`with_capacity` is a performance optimization that can turn O(n^2)
memory allocation patterns into O(1). Consider the following (toy) hot loop:

```rust
let mut vec = Vec::new();
for i in 0..100_000_000 {
  vec.push(i);
}
```

On each iteration, this will add a new element to the vector.
`Vec::new` in Rust creates a buffer with capacity 0, so this will start
at 0 and gradually increase the capacity as needed to hold all the
elements. Since the standard library doesn't know how many elements you
plan to allocate, it will double the capacity each time, for a total of
ceil(log_2(100_000_000)) = 27 reallocations. Each reallocation requires
copying the entire buffer - the last one will require copying 67 MB.

On my machine this runs in 148 ms:

```
$ hyperfine ./allocate --warmup 1
Benchmark #1: ./allocate
  Time (mean ± σ):     147.6 ms ±   0.5 ms    [User: 121.9 ms, System: 25.8 ms]
  Range (min … max):   146.7 ms … 148.7 ms    20 runs
```

This exponential growth can be avoided by pre-allocating the vec:

```
let mut vec = Vec::with_capacity(100_000_000);
for i in 0..100_000_000 {
  vec.push(i);
}
```

This runs in only 140 ms:

```
$ hyperfine ./preallocate --warmup 1
Benchmark #1: ./preallocate
  Time (mean ± σ):     139.8 ms ±   0.5 ms    [User: 121.1 ms, System: 18.8 ms]
  Range (min … max):   139.1 ms … 140.8 ms    21 runs
```

However, YDBRust does not have any such degenerate case where it
continually allocates to the vector (if it gets YDB_INVSTRLEN, it knows
exactly the capacity to expand to and will not need to expand again). So there is no need for pre-allocating any of the buffers.