[#31] Pre-allocate memory for new subscripts
If new subscripts are being allocated, that means that they're about to
be filled. Currently this will require calling into YDB once for each
subscript, because each call will return a separate YDB_INVSTRLEN.
This should allocate enough memory in the common case that INVSTRLEN
won't be return for most subscripts.