[#34] Fix infinite loop with multiple buffers and insufficient capacity.
In some rare cases, there would be multiple buffers passed to YDB (e.g.
in a `ydb_ci_t` call). However, the YDBRust wrapper previously assumed
there would only ever be one. This fixes the wrapper so that it will not
try to resize the buffer indefinitely if there is more than one buffer
passed; see the comments for more details.

Additionally, this adds a debug assertion that there is no infinite loop
and the buffer is resized only a fixed number of times.

- Use YDB_MAX_STR to be sure any INVSTRLEN is from the out_buffer
- Test and add comments around INVSTRLEN for node_next and sub_next

  These had no such bug. But the reason they didn't is non-trivial, so
  this adds a test and some comments as to why.