[YottaDB/DB/YDB!863] Update YDBRust for new version of `verify_commit.sh` in YDB
- Add `needs_copyright.sh`
- Change `pre-commit` to use `needs_copyright.sh`
- Change `.gitlab-ci.yml` to use the new name `verify_commit.sh` instead of `verify_gpg_commit.sh`

This also changes the pre-commit script to use `--no-default-features`;
this greatly decreases compile times when bindgen is installed. When
bindgen is not installed it will give an error message saying to install
it:

```
  --- stderr
  thread 'main' panicked at 'error: failed to run bindgen: No such file or directory (os error 2)
  help: make sure `bindgen` is installed and in $PATH
  help: consider enabling the `vendor` feature, which compiles `bindgen` from source', build.rs:85:21
  note: run with `RUST_BACKTRACE=1` environment variable to display a backtrace
```

Finally, this adds a copyright to the `.ci` call-in files.
Since https://gitlab.com/YottaDB/DB/YDB/-/issues/566 was implemented,
.ci files can now have comments.