Fix broken link in documentation
This fixes the following error from `cargo deadlinks`:

```
Found invalid urls in context_api/struct.Context.html:
	Fragment #intrinsic-variables at /home/joshua/.local/lib/cargo/target/doc/yottadb/context_api/index.html does not exist!
```

The link was always broken, but only recently was detected by deadlinks.
The similar related error was a bug in `deadlinks` itself:

```
Found invalid urls in craw/index.html:
	Fragment # at /home/joshua/.local/lib/cargo/target/doc/yottadb/craw/index.html does not exist!
```

It was fixed in the [0.6.1 release], so the pipeline should pass now.

[0.6.1 release]: https://github.com/deadlinks/cargo-deadlinks/blob/master/CHANGELOG.md#fixed