Use a `Cell` for TpToken instead of a `RefCell`
This has less performance overhead at runtime, because `Cell` is just a
normal pointer without `noalias`, while `RefCell` has internal state to
keep track of the number of borrows currently active.

This still has to use an Rc so that the same tptoken will be shared
between all `Context`s.