Fix comment on ydb_lock
It was incorrect - `args` can be either stack- or heap-allocated.

This also adds a compiler warning when compiling for 32-bit big-endian
platforms, which are untested and may have undefined behavior, as
noted in the documentation for `lock_st`. It looks like this:

```
$ PKG_CONFIG_ALLOW_CROSS=1 cargo check --target mips-unknown-linux-gnu
    Checking yottadb v1.2.0 (/home/joshua/src/yottadb/YDBRust)
warning: unused variable: `LOCK_ST_IS_UNTESTED_AND_UNSUPPORTED_ON_THIS_PLATFORM`
    --> src/simple_api/mod.rs:1924:17
     |
1924 |             let LOCK_ST_IS_UNTESTED_AND_UNSUPPORTED_ON_THIS_PLATFORM = ();
     |                 ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^ help: if this is intentional, prefix it with an underscore: `_LOCK_ST_IS_UNTESTED_AND_UNSUPPORTED_ON_THIS_PLATFORM`
     |
     = note: `#[warn(unused_variables)]` on by default
```