[#43] Return a buffer in KeyContext::next_sub and KeyContext::prev_sub
Previously, it returned a full key, which required a copy of the
variable and subscript. This can be easily rewritten in user code, so
this only returns the info needed to do it outside of YDBRust instead.

Since this is a breaking change, it also bumps the version to 2.0 to
avoid accidentally making a release that's marked as semver-compatible.