Remove unnecessary `out_buffer` param to `zwr2str`
This can just use the buffer of the Context, there's no need for it to
require a buffer from the user.

This also changes `str2zwr` to use the buffer of the context instead of
creating a new one, for consistency. The only difference is that this
can save an allocation if the user never uses `Context` after calling
str2zwr.