Revert "[Pipeline] Fix TTY issue with docker"
This reverts commit b10f119c5bdf87151dc0a97cbc6694c4a68a41b6.

It turns out the problem had to do with our pipeline configuration not
allowing volumes to be passed; and that warning message I was attempting
to fix was there all along, and is only a warning. The pipeline
configuration has now been fixed, and this no-op change is being
reverted.