Update bindgen to the latest version
On Ubuntu 23.10 and OpenSuse Tumbleweed, bindgen 0.59 failed to generate constants for anonymous enums:
```
error[E0432]: unresolved imports `crate::craw::YDB_DEL_TREE`, `crate::craw::YDB_DEL_NODE`
  --> src/simple_api/mod.rs:39:80
   |
39 |     ydb_tp_st, YDB_OK, YDB_ERR_INVSTRLEN, YDB_ERR_INSUFFSUBS, YDB_ERR_TPRETRY, YDB_DEL_TREE,
   |                                                                                ^^^^^^^^^^^^ no `YDB_DEL_TREE` in `craw`
40 |     YDB_DEL_NODE, YDB_TP_RESTART, YDB_TP_ROLLBACK,
   |     ^^^^^^^^^^^^ no `YDB_DEL_NODE` in `craw`
```

I have not fully dug into the root cause, but I suspect there was some change in the libclang API
between clang 13 (the default on Ubuntu 22.02) and clang 16 (the default on 23.10). Upgrading bindgen fixes the issue.

Note that at one point there was a theory that the version of the rust compiler was causing the
issue. I reproduced this issue on Ubuntu 23.10 with Rust 1.67, so I don't think the toolchain is
related.